added an auth redirect, removed unused code, fixed routes, refactored home to a folder, removed unused styling from Layout.js
