+++ client/src/features/layout/Layout.js
@@ -1,32 +1,21 @@
 import React from 'react'
-import { makeStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import IconButton from '@material-ui/core/IconButton';
 // import MenuIcon from '@material-ui/icons/Menu';
 
-const useStyles = makeStyles(theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-}));
-
 export default function Layout() {
-  const classes = useStyles();
+
 
   return (
-    <div className={classes.root}>
-      <AppBar position="static">
-        <Toolbar variant="dense">
-          <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
-            {/* <MenuIcon /> */}
+    <div >
+      <AppBar>
+        <Toolbar>
+          <IconButton>
           </IconButton>
-          <Typography variant="h6" color="inherit">
-            Menu Title
+          <Typography variant="h6">
+            Clark Real Estate
           </Typography>
         </Toolbar>
       </AppBar>
